Add BuildList component and display builds if available
This commit adds a new file 'build-list.tsx' that contains the `BuildList` component. It takes an array of builds and renders them as a list. The `MyBuilds` component now includes the `BuildList` component, passing it the array of builds to display. If there are no builds available, a message indicating so is shown.