feat: Add isDevelopment utility
This commit adds a new utility function `isDevelopment` to the `utils` folder. It checks if the current environment is development based on the `process.env.NODE_ENV` value. This utility is used in the `Nineball` component to conditionally execute the code block that appends a message. The utility is also imported and used in the `nineball.tsx` file.